reduce ram usage, use u16

+++ inclusion_exclusion/prototype_solver/src/clauses/mod.rs
@@ -17,7 +17,7 @@ where
     Self: Sized + Clone + Default + Debug,
 {
     fn merge(a: Self, b: &Self) -> MergeResult<Self>;
-    fn from_vec(vec: Vec<Vec<(u32, Sign)>>) -> DNF<Self>;
+    fn from_vec(vec: Vec<Vec<(u16, Sign)>>) -> DNF<Self>;
     fn len(&self) -> usize;
     fn new_empty() -> Self;
 }
